Refactor structure for feedback components.
